fix: Corrects the img tag sizes attribute unit
The img tag sizes attribute used % instead of the acceptable unit such
as vw. Additionally, the media query had to be in reverse order. These
issues were detected by the HTML validator.

My original intent of the sizes value were:
* Default to 33% of the container width
* At screen size less than 640px, set to 50% of the container width
* At screen size less than 480px set to 100% of the container width

I finally noticed that the sizes attribute only applies to the entire
view port width, not to the container. Additionally, the media query had
to be reversed to work correctly.

The updated sizes value now says:
* At screen size of at least 480px, image size is 38vw (derived by the
grid width of 75vw * column width of 50%)
* At screen size of at least 640px, image size is 25vw (derived by the
grid width of 75vw * column width of 33.33%)
* By default (screen size of less than 480px), the image size is the
same as the grid width, which is 75vw.